Remove left over console commands.
Accidently left a clear console command in the standard write command.
This would result in screen clearing text during a dump operation if the
dump target was the DSX cart. This might make the user believe it has
hung when infact it has not. So I have corrected this. :P