The big release overhaul...
* Option to write a default fat image to card is now possible through
new covery menu! An image for 16g capacity cards is included. Note that
other capacities are not currently supported until fat image jumps of
them become available.
* Recovery menu has option to dump/write sector 0 of FAT section of
nand. This allows restoring bricked cards if access to a DS/DS Lite and
the reguired slot2 device is not available.
* Ability to boot into bootstrap program to launch files on cart added
to recovery menu. Allows booting udisk updator as a result so can
recover a bricked cart!
* Changes to UI. New top screen graphic. Graphics for bottom screen now
added for better console! :D