Optimize Bootstrap, add check on init command...
* Optimized Bootstrap. Should transition quicker when boot into it as
well as into what ever the bootstrap boots.
* New check added for initial data return by the first special cart init
command. If the data returned in the buffer used for the init command
doesn't match the known value, a warning message is shown and the init
data is logged to bin file. User can choose to ignore the warning by
pressing A.