1.3 Release Commit...
* Revamped card read code to use D2 command documented from old source
code to DLDI driver. This appears to result in proper dump of hidden
nand sections. As a result the "alt test dump" mode has been removed as
the data it would have returned shows up in the main test dump file now.

* Some more tweaks to UI text.

* NitroFS implemented. Basic fat images and bootstrap srl now available
via NitroFS is nrioFolder on SD card doesn't contain the needed files.

* Fat image restore feature now enforces strict capacity requirements.
The included images are intended for carts with a specific nand
capacity. If they don't match it will not use them.
* Option to dump full fat filesystem image added to DLDI menu.
* Option to write nrio_customimage.img added as override to fat image
restore feature. Use this if you wish to restore a dump or want to
override the capacity requirement (though doing so is not recommended.
You can build your own via a dump if you have a functional fat
filesystem on the cart already)
* optional auto boot setting added to xBootstrap (XuluMenu) code. If
compiled with autoboot set on, XuluMenu can act as a xmenu replacement
that will autoboot xmenu.srl unless B or Y buttons are held on boot.
Holding B on boot would bring up file browser while Y will auto boot
udisk.nds on root of nand if present.